# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ZICK — A spiritual-wellness app backend providing auth, content, schedules, and subscriptions.",
        "title": "ZICK API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ZICK API Support",
            "email": "support@zick.app"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Uploads any file (image, audio, etc.) to Cloudinary and returns the public URL.\nPass an optional `folder` query parameter to organise assets (e.g. `?folder=content/thumbnails`).\nThe returned `url` can be stored and referenced by any domain (content, profile, etc.).",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "Upload a file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cloudinary folder (default: zick/uploads)",
                        "name": "folder",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_media.UploadResponse"
                        }
                    },
                    "400": {
                        "description": "Missing or oversized file",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "503": {
                        "description": "Upload service unavailable",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently removes a file from Cloudinary using its public_id.\nThe public_id is returned by POST /api/v1/upload.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Media"
                ],
                "summary": "Delete an uploaded file",
                "parameters": [
                    {
                        "description": "Full Cloudinary URL to delete",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_domain_media.DeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_domain_media.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Missing public_id",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    },
                    "503": {
                        "description": "Upload service unavailable",
                        "schema": {
                            "$ref": "#/definitions/gotickets_internal_httpresponse.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gotickets_internal_httpresponse.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_domain_media.DeleteRequest": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "url": {
                    "type": "string",
                    "example": "https://res.cloudinary.com/demo/image/upload/v1234/zick/uploads/abc.jpg"
                }
            }
        },
        "internal_domain_media.DeleteResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "File deleted successfully"
                }
            }
        },
        "internal_domain_media.UploadResponse": {
            "type": "object",
            "properties": {
                "public_id": {
                    "type": "string",
                    "example": "zick/content/abc123"
                },
                "url": {
                    "type": "string",
                    "example": "https://res.cloudinary.com/demo/image/upload/sample.jpg"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}